+++ src/OrchardCore/OrchardCore.Abstractions/Extensions/JsonSerializerOptionsConfiguration.cs
@@ -28,6 +28,5 @@ public class JsonSerializerOptionsConfiguration : IConfigureOptions<JsonSerializ
         options.TypeInfoResolverChain.Add(new PolymorphicJsonTypeInfoResolver(_derivedTypesOptions));
         options.Converters.Add(DynamicJsonConverter.Instance);
         options.Converters.Add(PathStringJsonConverter.Instance);
-        options.Converters.Add(LoginInfoJsonConverter.Instance);
     }
 }
