+++ src/OrchardCore/OrchardCore.Abstractions/Json/JOptions.cs
@@ -31,8 +31,6 @@ public static class JOptions
 
     static JOptions()
     {
-        Base.Converters.Add(new LoginInfoJsonConverter());
-
         Default = new JsonSerializerOptions(Base);
         Default.Converters.Add(new DynamicJsonConverter());
         Default.Converters.Add(new PathStringJsonConverter());
