+++ src/OrchardCore/OrchardCore.Users.Core/Extensions/UsersServiceCollectionExtensions.cs
@@ -9,6 +9,7 @@ using OrchardCore.Users.Events;
 using OrchardCore.Users.Handlers;
 using OrchardCore.Users.Indexes;
 using OrchardCore.Users.Services;
+using OrchardCore.Users.Core.Json;
 
 namespace Microsoft.Extensions.DependencyInjection
 {
@@ -54,6 +55,11 @@ namespace Microsoft.Extensions.DependencyInjection
 
             services.AddScoped<ITwoFactorAuthenticationHandler, DefaultTwoFactorAuthenticationHandler>();
             services.AddScoped<ITwoFactorAuthenticationHandlerCoordinator, DefaultTwoFactorAuthenticationHandlerCoordinator>();
+
+            services.Configure<JsonSerializerOptions>(options =>
+            {
+                options.Converters.Add(new LoginInfoJsonConverter())
+            });
 
             return services;
         }
